Struct basic_dsp::combined_ops::Identifier

pub struct Identifier<T, N, D> where
    D: Domain,
    N: NumberSpace,
    T: RealNumber,  { /* fields omitted */ }

An identifier is just a placeholder for a data type used to ensure already at compile time that operations are valid.