Struct basic_dsp::combined_ops::MultiOperation1

pub struct MultiOperation1<T, TO> where T: RealNumber, TO: ToIdentifier<T> {
    // some fields omitted
}

A multi operation which holds a vector and records all changes which need to be done to the vectors. By calling get on the struct all operations will be executed in one run.

Methods

impl<TO> MultiOperation1<f32, TO> where TO: ToIdentifier<f32> + DataVector<f32> + RededicateVector<GenericDataVector<f32>>

Holds a vector and records all operations which shall be done on the vector. A call to get then runs all recorded operations on the vector and returns it. See the modules description for why this can be beneficial.

fn get(self) -> Result<TO, (ErrorReason, TO)>

Executes all recorded operations on the stored vectors.

fn add_ops<F, TN>(self, operation: F) -> MultiOperation1<f32, TN::Vector> where F: Fn(TO::Identifier) -> TN, TN: Identifier<f32>

Adds new operations which will be executed with the next call to get

As a background: The function operation will be executed immediately. It only operated on Identifier types and these serve as placeholder for vectors. Every operation done to an Identifier is recorded and will be executed on vectors if get is called.

fn extend<TI2>(self, vector: TI2) -> MultiOperation2<f32, TO, TI2> where TI2: ToIdentifier<f32> + DataVector<f32> + RededicateVector<GenericDataVector<f32>>

Extends the operation to operate on one more vector.

impl<TO> MultiOperation1<f64, TO> where TO: ToIdentifier<f64> + DataVector<f64> + RededicateVector<GenericDataVector<f64>>

Holds a vector and records all operations which shall be done on the vector. A call to get then runs all recorded operations on the vector and returns it. See the modules description for why this can be beneficial.

fn get(self) -> Result<TO, (ErrorReason, TO)>

Executes all recorded operations on the stored vectors.

fn add_ops<F, TN>(self, operation: F) -> MultiOperation1<f64, TN::Vector> where F: Fn(TO::Identifier) -> TN, TN: Identifier<f64>

Adds new operations which will be executed with the next call to get

As a background: The function operation will be executed immediately. It only operated on Identifier types and these serve as placeholder for vectors. Every operation done to an Identifier is recorded and will be executed on vectors if get is called.

fn extend<TI2>(self, vector: TI2) -> MultiOperation2<f64, TO, TI2> where TI2: ToIdentifier<f64> + DataVector<f64> + RededicateVector<GenericDataVector<f64>>

Extends the operation to operate on one more vector.

impl<T> MultiOperation1<T, GenericDataVector<T>> where T: RealNumber

fn add_enum_op(&mut self, op: Operation<T>)

Allows to directly push an Operation enum to a MultiOperation1. This mainly exists as interop between Rust and other languages.