Struct basic_dsp::combined_ops::PreparedOperation1

pub struct PreparedOperation1<T, TI1, TO1> where T: RealNumber, TI1: ToIdentifier<T>, TO1: ToIdentifier<T> {
    // some fields omitted
}

An operation on one data vector which has been prepared in advance.

The type arguments can be read like a function definition: TI1 -> TO1

Methods

impl<TI1, TO1> PreparedOperation1<f32, TI1, TO1> where TI1: ToIdentifier<f32> + DataVector<f32> + RededicateVector<GenericDataVector<f32>>, TO1: ToIdentifier<f32> + DataVector<f32> + RededicateVector<GenericDataVector<f32>>

An operation which can be prepared in advance and operates on one input and produces one output

fn extend<TI2>(self) -> PreparedOperation2<f32, TI1, TI2, TO1, TI2> where TI2: ToIdentifier<f32>

Extends the operation to operate on one more vector.

fn add_ops<F, TN>(self, operation: F) -> PreparedOperation1<f32, TI1, TN::Vector> where F: Fn(TO1::Identifier) -> TN, TN: Identifier<f32>

Adds new operations which will be executed with the next call to exec

As a background: The function operation will be executed immediatly. It only operated on Identifier types and these serve as placeholder for vectors. Every operation done to an Identifier is recorded and will be executed on vectors if exec is called.

fn exec(&self, a: TI1) -> Result<TO1, (ErrorReason, TO1)>

Executes all recorded operations on the input vectors.

impl<TI1, TO1> PreparedOperation1<f64, TI1, TO1> where TI1: ToIdentifier<f64> + DataVector<f64> + RededicateVector<GenericDataVector<f64>>, TO1: ToIdentifier<f64> + DataVector<f64> + RededicateVector<GenericDataVector<f64>>

An operation which can be prepared in advance and operates on one input and produces one output

fn extend<TI2>(self) -> PreparedOperation2<f64, TI1, TI2, TO1, TI2> where TI2: ToIdentifier<f64>

Extends the operation to operate on one more vector.

fn add_ops<F, TN>(self, operation: F) -> PreparedOperation1<f64, TI1, TN::Vector> where F: Fn(TO1::Identifier) -> TN, TN: Identifier<f64>

Adds new operations which will be executed with the next call to exec

As a background: The function operation will be executed immediatly. It only operated on Identifier types and these serve as placeholder for vectors. Every operation done to an Identifier is recorded and will be executed on vectors if exec is called.

fn exec(&self, a: TI1) -> Result<TO1, (ErrorReason, TO1)>

Executes all recorded operations on the input vectors.

impl<T> PreparedOperation1<T, GenericDataVector<T>, GenericDataVector<T>> where T: RealNumber

fn add_enum_op(&mut self, op: Operation<T>)

Allows to directly push an Operation enum to a PreparedOperation1. This mainly exists as interop between Rust and other languages.