Trait basic_dsp::combined_ops::ComplexIdentifier

pub trait ComplexIdentifier<T>: Identifier<T> where T: RealNumber {
    type RealPartner;
    fn complex_offset(self, offset: Complex<T>) -> Self;
    fn complex_scale(self, factor: Complex<T>) -> Self;
    fn magnitude(self) -> Self::RealPartner;
    fn magnitude_squared(self) -> Self::RealPartner;
    fn conj(self) -> Self;
    fn to_real(self) -> Self::RealPartner;
    fn to_imag(self) -> Self::RealPartner;
    fn phase(self) -> Self::RealPartner;
    fn multiply_complex_exponential(self, a: T, b: T) -> Self;
    fn map_inplace_complex<F>(self, f: F) -> Self where F: Fn(Complex<T>, usize) -> Complex<T> + Send + Sync;
}

Operations for complex vectors which can be used in combination with multi ops or prepared ops. For a documentation of the specific operations see ComplexVectorOps.

Associated Types

type RealPartner

Required Methods

fn complex_offset(self, offset: Complex<T>) -> Self

See ComplexVectorOps.

fn complex_scale(self, factor: Complex<T>) -> Self

See ComplexVectorOps.

fn magnitude(self) -> Self::RealPartner

See ComplexVectorOps.

fn magnitude_squared(self) -> Self::RealPartner

See ComplexVectorOps.

fn conj(self) -> Self

See ComplexVectorOps.

fn to_real(self) -> Self::RealPartner

See ComplexVectorOps.

fn to_imag(self) -> Self::RealPartner

See ComplexVectorOps.

fn phase(self) -> Self::RealPartner

See ComplexVectorOps.

fn multiply_complex_exponential(self, a: T, b: T) -> Self

See ComplexVectorOps.

fn map_inplace_complex<F>(self, f: F) -> Self where F: Fn(Complex<T>, usize) -> Complex<T> + Send + Sync

See ComplexVectorOps.