Trait basic_dsp::combined_ops::GeneralIdentifier

pub trait GeneralIdentifier<T>: Identifier<T> where T: RealNumber {
    fn add_vector(self, summand: &Self) -> Self;
    fn subtract_vector(self, subtrahend: &Self) -> Self;
    fn multiply_vector(self, factor: &Self) -> Self;
    fn divide_vector(self, divisor: &Self) -> Self;
    fn sqrt(self) -> Self;
    fn square(self) -> Self;
    fn root(self, degree: T) -> Self;
    fn powf(self, exponent: T) -> Self;
    fn ln(self) -> Self;
    fn exp(self) -> Self;
    fn log(self, base: T) -> Self;
    fn expf(self, base: T) -> Self;
    fn sin(self) -> Self;
    fn cos(self) -> Self;
    fn tan(self) -> Self;
    fn asin(self) -> Self;
    fn acos(self) -> Self;
    fn atan(self) -> Self;
    fn sinh(self) -> Self;
    fn cosh(self) -> Self;
    fn tanh(self) -> Self;
    fn asinh(self) -> Self;
    fn acosh(self) -> Self;
    fn atanh(self) -> Self;
    fn clone_from(self, &Self) -> Self;
    fn add_points(self) -> Self;
    fn sub_points(self) -> Self;
    fn div_points(self) -> Self;
    fn mul_points(self) -> Self;
}

Operations for all kind of vectors which can be used in combination with multi ops or prepared ops. For a documentation of the specific operations see GenericVectorOps.

Required Methods

fn add_vector(self, summand: &Self) -> Self

See GenericVectorOps.

fn subtract_vector(self, subtrahend: &Self) -> Self

See GenericVectorOps.

fn multiply_vector(self, factor: &Self) -> Self

See GenericVectorOps.

fn divide_vector(self, divisor: &Self) -> Self

See GenericVectorOps.

fn sqrt(self) -> Self

See GenericVectorOps.

fn square(self) -> Self

See GenericVectorOps.

fn root(self, degree: T) -> Self

See GenericVectorOps.

fn powf(self, exponent: T) -> Self

See GenericVectorOps.

fn ln(self) -> Self

See GenericVectorOps.

fn exp(self) -> Self

See GenericVectorOps.

fn log(self, base: T) -> Self

See GenericVectorOps.;

fn expf(self, base: T) -> Self

See GenericVectorOps.

fn sin(self) -> Self

See GenericVectorOps.

fn cos(self) -> Self

See GenericVectorOps.

fn tan(self) -> Self

See GenericVectorOps.

fn asin(self) -> Self

See GenericVectorOps.

fn acos(self) -> Self

See GenericVectorOps.

fn atan(self) -> Self

See GenericVectorOps.

fn sinh(self) -> Self

See GenericVectorOps.

fn cosh(self) -> Self

See GenericVectorOps.

fn tanh(self) -> Self

See GenericVectorOps.

fn asinh(self) -> Self

See GenericVectorOps.

fn acosh(self) -> Self

See GenericVectorOps.

fn atanh(self) -> Self

See GenericVectorOps.

fn clone_from(self, &Self) -> Self

Copies data from another vector.

fn add_points(self) -> Self

Adds its length to the vector elements

Example

use basic_dsp::*;
use basic_dsp::combined_ops::*;    
let complex = ComplexTimeVector32::from_interleaved(&[1.0, 2.0, 3.0, 4.0]);
let ops = multi_ops1(complex);
let ops = ops.add_ops(|v|v.add_points());
let complex = ops.get().expect("Ignoring error handling in examples");
assert_eq!([3.0, 2.0, 5.0, 4.0], complex.data());

fn sub_points(self) -> Self

Subtracts its length from the vector elements

Example

use basic_dsp::*;
use basic_dsp::combined_ops::*;    
let complex = ComplexTimeVector32::from_interleaved(&[3.0, 2.0, 5.0, 4.0]);
let ops = multi_ops1(complex);
let ops = ops.add_ops(|v|v.sub_points());
let complex = ops.get().expect("Ignoring error handling in examples");
assert_eq!([1.0, 2.0, 3.0, 4.0], complex.data());

fn div_points(self) -> Self

divides the vector elements by its length Subtracts its length from the vector elements

Example

use basic_dsp::*;
use basic_dsp::combined_ops::*;    
let complex = ComplexTimeVector32::from_interleaved(&[2.0, 4.0, 6.0, 8.0]);
let ops = multi_ops1(complex);
let ops = ops.add_ops(|v|v.div_points());
let complex = ops.get().expect("Ignoring error handling in examples");
assert_eq!([1.0, 2.0, 3.0, 4.0], complex.data());

fn mul_points(self) -> Self

Multiplies the vector elements with its length

Example

use basic_dsp::*;
use basic_dsp::combined_ops::*;    
let complex = ComplexTimeVector32::from_interleaved(&[1.0, 2.0, 3.0, 4.0]);
let ops = multi_ops1(complex);
let ops = ops.add_ops(|v|v.mul_points());
let complex = ops.get().expect("Ignoring error handling in examples");
assert_eq!([2.0, 4.0, 6.0, 8.0], complex.data());