Struct basic_dsp::conv_types::ComplexFrequencyLinearTableLookup

pub struct ComplexFrequencyLinearTableLookup<T> where T: RealNumber {
    // some fields omitted
}

Allows to create a lookup table with linear interpolation between table points. This usually speeds up a convolution and sacrifies accuracy.

Methods

impl<T> ComplexFrequencyLinearTableLookup<T> where T: RealNumber

fn table(&self) -> &[Complex<T>]

Allows to inspect the generated lookup table

fn delta(&self) -> T

Gets the delta value which determines the resolution

impl ComplexFrequencyLinearTableLookup<f32>

fn from_raw_parts(table: &[Complex32], delta: f32, is_symmetric: bool) -> Self

Creates a lookup table by putting the pieces together.

fn from_conv_function(other: &ComplexFrequencyResponse<f32>, delta: f32, len: usize) -> Self

Creates a lookup table from another convolution function. The delta argument can be used to balance performance vs. accuracy.

impl ComplexFrequencyLinearTableLookup<f64>

fn from_raw_parts(table: &[Complex64], delta: f64, is_symmetric: bool) -> Self

Creates a lookup table by putting the pieces together.

fn from_conv_function(other: &ComplexFrequencyResponse<f64>, delta: f64, len: usize) -> Self

Creates a lookup table from another convolution function. The delta argument can be used to balance performance vs. accuracy.

impl ComplexFrequencyLinearTableLookup<f32>

fn to_real(self) -> RealFrequencyLinearTableLookup<f32>

Convert the lookup table into real number space

impl ComplexFrequencyLinearTableLookup<f64>

fn to_real(self) -> RealFrequencyLinearTableLookup<f64>

Convert the lookup table into real number space

impl ComplexFrequencyLinearTableLookup<f32>

fn ifft(self) -> ComplexTimeLinearTableLookup<f32>

Convert the lookup table into time domain

impl ComplexFrequencyLinearTableLookup<f64>

fn ifft(self) -> ComplexTimeLinearTableLookup<f64>

Convert the lookup table into time domain