Struct basic_dsp::conv_types::ForeignComplexConvolutionFunction

pub struct ForeignComplexConvolutionFunction<T> where T: RealNumber {
    pub conv_function:  extern "C" fn(*const c_void, T) -> Complex<T>,
    pub conv_data: usize,
    pub is_symmetric: bool,
}

A complex function which can be constructed outside this crate.

Fields

conv_function

The function

conv_data

The data which is passed to the window function

Actual data type is a const* c_void, but Rust doesn't allow that becaues it's usafe so we store it as usize and transmute it when necessary. Callers shoulds make very sure safety is guaranteed.

is_symmetric

Indicates whether this function is symmetric around 0 or not. Symmetry is defined as self.calc(x) == self.calc(-x).

Methods

impl<T> ForeignComplexConvolutionFunction<T> where T: RealNumber

fn new(function: extern "C" fn(*const c_void, T) -> Complex<T>, function_data: *const c_void, is_symmetric: bool) -> Self

Creates a new real function