Struct basic_dsp::conv_types::RealTimeLinearTableLookup

pub struct RealTimeLinearTableLookup<T> where T: RealNumber {
    // some fields omitted
}

Allows to create a lookup table with linear interpolation between table points. This usually speeds up a convolution and sacrifies accuracy.

Methods

impl<T> RealTimeLinearTableLookup<T> where T: RealNumber

fn table(&self) -> &[T]

Allows to inspect the generated lookup table

fn delta(&self) -> T

Gets the delta value which determines the resolution

impl RealTimeLinearTableLookup<f32>

fn from_raw_parts(table: &[f32], delta: f32, is_symmetric: bool) -> Self

Creates a lookup table by putting the pieces together.

fn from_conv_function(other: &RealImpulseResponse<f32>, delta: f32, len: usize) -> Self

Creates a lookup table from another convolution function. The delta argument can be used to balance performance vs. accuracy.

impl RealTimeLinearTableLookup<f64>

fn from_raw_parts(table: &[f64], delta: f64, is_symmetric: bool) -> Self

Creates a lookup table by putting the pieces together.

fn from_conv_function(other: &RealImpulseResponse<f64>, delta: f64, len: usize) -> Self

Creates a lookup table from another convolution function. The delta argument can be used to balance performance vs. accuracy.

impl RealTimeLinearTableLookup<f32>

fn to_complex(&self) -> ComplexTimeLinearTableLookup<f32>

Convert the lookup table into complex number space

impl RealTimeLinearTableLookup<f64>

fn to_complex(&self) -> ComplexTimeLinearTableLookup<f64>

Convert the lookup table into complex number space

impl RealTimeLinearTableLookup<f32>

fn fft(self) -> RealFrequencyLinearTableLookup<f32>

Convert the lookup table into a magnitude spectrum

impl RealTimeLinearTableLookup<f64>

fn fft(self) -> RealFrequencyLinearTableLookup<f64>

Convert the lookup table into a magnitude spectrum