Enum basic_dsp::Operation

pub enum Operation<T> {
    AddReal(T),
    AddComplex(Complex<T>),
    MultiplyReal(T),
    MultiplyComplex(Complex<T>),
    AbsReal,
    AbsComplex,
    Sqrt,
}

An alternative way to define operations on a vector. Warning: Highly unstable and not even fully implemented right now.

In future this enum will likely be deleted or hidden and be replaced with a builder pattern. The advantage of this is that with the builder we have the means to define at compile time what kind of vector will result from the given set of operations.

Variants

AddReal

AddComplex

MultiplyReal

MultiplyComplex

AbsReal

AbsComplex

Sqrt