Module basic_dsp::interop_facade

Clients using other programming languages should use the functions in this mod. Please refer to the other chapters of the help for documentation of the functions.

Modules

facade32
Functions for 32bit floating point number based vectors. Please refer to the other chapters of the help for documentation of the functions.

facade64
Functions for 64bit floating point number based vectors. Please refer to the other chapters of the help for documentation of the functions.

Structs

BinaryVectorResult
Result of a vector operation. Check the result_code.

ScalarResult
Result of a vector operation. Check the result_code.

VectorResult
Result of a vector operation. Check the result_code.

Functions

translate_error
Error codes:

translate_to_padding_option

translate_to_real_convolution_function

translate_to_real_frequency_response

translate_to_window_function