Struct basic_dsp::interop_facade::VectorResult

pub struct VectorResult<T> {
    pub result_code: i32,
    pub vector: Box<T>,
}

Result of a vector operation. Check the result_code.

Fields
result_code: i32

This value is zero in case of error. All other values mean that an error occurred and the data in the vector might be unchanged or invalid. Error codes are described in translate_error.

vector: Box<T>

A pointer to a data vector.