Trait basic_dsp::matrix::ToDspMatrix

pub trait ToDspMatrix<V, T> where
    T: RealNumber,
    V: Vector<T>, {
    type Output: Matrix<V, T>;
    fn to_gen_dsp_mat(
        self, 
        is_complex: bool, 
        domain: DataDomain
    ) -> Self::Output;
}

Conversion from a generic data type into a dsp matrix which tracks its meta information (domain and number space) only at runtime. See ToRealMatrix and ToComplexMatrix for alternatives which track most of the meta data with the type system and therefore avoid runtime errors.

Associated Types

type Output: Matrix<V, T>

Loading content...
Required methods

fn to_gen_dsp_mat(self, is_complex: bool, domain: DataDomain) -> Self::Output

Create a new generic matrix. delta can be changed after construction with a call of set_delta.

For complex matrices with an odd length the resulting value will have a zero length.

Loading content...