Trait basic_dsp::matrix::ToRealTimeMatrix

pub trait ToRealTimeMatrix<V, T> where
    T: RealNumber,
    V: Vector<T>, {
    type Output: Matrix<V, T>;
    fn to_real_time_mat(self) -> Self::Output;
}

Conversion from a generic data type into a dsp matrix with real data.

Associated Types

type Output: Matrix<V, T>

Loading content...
Required methods

fn to_real_time_mat(self) -> Self::Output

Create a new matrix in real number space and time domain. delta can be changed after construction with a call of set_delta.

Loading content...