Struct basic_dsp::ComplexFreqVector

pub struct ComplexFreqVector<T> where T: RealNumber {
    // some fields omitted
}

A 1xN (one times N elements) or Nx1 data vector as used for most digital signal processing (DSP) operations. All data vector operations consume the vector they operate on and return a new vector. A consumed vector must not be accessed again.

Vectors come in different flavors:

Time or Frequency domain

Real or Complex numbers

32bit or 64bit floating point numbers

The first two flavors define meta information about the vector and provide compile time information what operations are available with the given vector and how this will transform the vector. This makes sure that some invalid operations are already discovered at compile time. In case that this isn't desired or the information about the vector isn't known at compile time there are the generic DataVector32 and DataVector64 vectors available.

32bit and 64bit flavors trade performance and memory consumption against accuracy. 32bit vectors are roughly two times faster than 64bit vectors for most operations. But remember that you should benchmark first before you give away accuracy for performance unless however you are sure that 32bit accuracy is certainly good enough.

Methods

impl<T> ComplexFreqVector<T> where T: RealNumber

fn from_interleaved_no_copy_with_options(data: Vec<T>, options: MultiCoreSettings) -> Self

Same as from_interleaved_no_copy but also allows to set multicore options.

fn from_interleaved_with_options(data: &[T], options: MultiCoreSettings) -> Self

Same as from_interleaved but also allows to set multicore options.

fn from_interleaved_with_delta_and_options(data: &[T], delta: T, options: MultiCoreSettings) -> Self

Same as from_interleaved_with_delta but also allows to set multicore options.

fn empty_with_options(options: MultiCoreSettings) -> Self

Same as complex_empty but also allows to set multicore options.

fn empty_with_delta_and_options(delta: T, options: MultiCoreSettings) -> Self

Same as complex_empty_with_delta but also allows to set multicore options.

fn from_constant_with_options(constant: T, length: usize, options: MultiCoreSettings) -> Self

Same as complex_from_constant but also allows to set multicore options.

fn from_constant_with_delta_and_options(constant: T, length: usize, delta: T, options: MultiCoreSettings) -> Self

Same as complex_from_constant_with_delta but also allows to set multicore options.

fn from_real_imag_with_options(real: &[T], imag: &[T], options: MultiCoreSettings) -> Self

Same as from_real_imag but also allows to set multicore options.

fn from_real_imag_with_delta_and_options(real: &[T], imag: &[T], delta: T, options: MultiCoreSettings) -> Self

Same as from_real_imag_with_delta but also allows to set multicore options.

fn from_mag_phase_with_options(magnitude: &[T], phase: &[T], options: MultiCoreSettings) -> Self

Same as from_mag_phase but also allows to set multicore options.

fn from_mag_phase_with_delta_and_options(magnitude: &[T], phase: &[T], delta: T, options: MultiCoreSettings) -> Self

Same as from_mag_phase_with_delta but also allows to set multicore options.

fn from_interleaved_no_copy(data: Vec<T>) -> Self

Creates a complex DataVector by consuming a Vec. Data is in interleaved format: i0, q0, i1, q1, ....

This operation is more memory efficient than the other options to create a vector, however if used outside of Rust then it holds the risk that the user will access the data parameter after the vector has been created causing all types of issues.

fn from_interleaved(data: &[T]) -> Self

Creates a complex DataVector from an array or sequence. Data is in interleaved format: i0, q0, i1, q1, .... delta is defaulted to 1.

fn from_interleaved_with_delta(data: &[T], delta: T) -> Self

Creates a complex DataVector from an array or sequence. Data is in interleaved format: i0, q0, i1, q1, .... delta is set to the given value.

fn empty() -> Self

Creates a complex and empty DataVector and sets delta to 1.0 value.

fn empty_with_delta(delta: T) -> Self

Creates a complex and empty DataVector and sets delta to the given value.

fn from_constant(constant: T, length: usize) -> Self

Creates a complex DataVector with length elements all set to the value of constant. delta is defaulted to 1.

fn from_constant_with_delta(constant: T, length: usize, delta: T) -> Self

Creates a complex DataVector with length elements all set to the value of constant and sets delta to the given value.

fn from_real_imag(real: &[T], imag: &[T]) -> Self

Creates a complex DataVector from an array with real and an array imaginary data. delta is set to 1.

Arrays must have the same length.

fn from_real_imag_with_delta(real: &[T], imag: &[T], delta: T) -> Self

Creates a complex DataVector from an array with real and an array imaginary data. delta is set to the given value.

Arrays must have the same length.

fn from_mag_phase(magnitude: &[T], phase: &[T]) -> Self

Creates a complex DataVector from an array with magnitude and an array with phase data. delta is set to 1.

Arrays must have the same length. Phase must be in [rad].

fn from_mag_phase_with_delta(magnitude: &[T], phase: &[T], delta: T) -> Self

Creates a complex DataVector from an array with magnitude and an array with phase data. delta is set to the given value.

Arrays must have the same length. Phase must be in [rad].

fn from_complex(complex: &[Complex<T>]) -> Self

Creates a complex DataVector from an array of complex numbers. delta is set to 1.

fn from_complex_with_delta(complex: &[Complex<T>], delta: T) -> Self

Creates a complex DataVector from an array of complex numbers. delta is set to the given value.

fn from_complex_with_delta_and_options(complex: &[Complex<T>], delta: T, options: MultiCoreSettings) -> Self

Creates a complex DataVector from an array of complex numbers.