Struct basic_dsp::GenericDataVector

pub struct GenericDataVector<T> where T: RealNumber {
    // some fields omitted
}

A 1xN (one times N elements) or Nx1 data vector as used for most digital signal processing (DSP) operations. All data vector operations consume the vector they operate on and return a new vector. A consumed vector must not be accessed again.

Vectors come in different flavors:

Time or Frequency domain

Real or Complex numbers

32bit or 64bit floating point numbers

The first two flavors define meta information about the vector and provide compile time information what operations are available with the given vector and how this will transform the vector. This makes sure that some invalid operations are already discovered at compile time. In case that this isn't desired or the information about the vector isn't known at compile time there are the generic DataVector32 and DataVector64 vectors available.

32bit and 64bit flavors trade performance and memory consumption against accuracy. 32bit vectors are roughly two times faster than 64bit vectors for most operations. But remember that you should benchmark first before you give away accuracy for performance unless however you are sure that 32bit accuracy is certainly good enough.

Methods

impl GenericDataVector<f32>

fn new_with_options(is_complex: bool, domain: DataVectorDomain, init_value: f32, length: usize, delta: f32, options: MultiCoreSettings) -> Self

Same as new but also allows to set multicore options.

fn from_array_with_options(is_complex: bool, domain: DataVectorDomain, data: &[f32], options: MultiCoreSettings) -> Self

Same as from_array but also allows to set multicore options.

fn from_array_no_copy_with_options(is_complex: bool, domain: DataVectorDomain, data: Vec<f32>, options: MultiCoreSettings) -> Self

Same as from_array_no_copy but also allows to set multicore options.

fn from_array_with_delta_and_options(is_complex: bool, domain: DataVectorDomain, data: &[f32], delta: f32, options: MultiCoreSettings) -> Self

Same as from_array_with_delta but also allows to set multicore options.

fn from_array_no_copy_with_delta_and_options(is_complex: bool, domain: DataVectorDomain, data: Vec<f32>, delta: f32, options: MultiCoreSettings) -> Self

Same as from_array_no_copy_with_delta but also allows to set multicore options.

fn new(is_complex: bool, domain: DataVectorDomain, init_value: f32, length: usize, delta: f32) -> Self

Creates a new generic data vector from the given arguments.

fn from_array(is_complex: bool, domain: DataVectorDomain, data: &[f32]) -> Self

Creates a new generic data vector from the given arguments.

fn from_array_no_copy(is_complex: bool, domain: DataVectorDomain, data: Vec<f32>) -> Self

Creates a new generic data vector from the given arguments.

fn from_array_with_delta(is_complex: bool, domain: DataVectorDomain, data: &[f32], delta: f32) -> Self

Creates a new generic data vector from the given arguments.

fn from_array_no_copy_with_delta(is_complex: bool, domain: DataVectorDomain, data: Vec<f32>, delta: f32) -> Self

Creates a new generic data vector from the given arguments.

impl GenericDataVector<f64>

fn new_with_options(is_complex: bool, domain: DataVectorDomain, init_value: f64, length: usize, delta: f64, options: MultiCoreSettings) -> Self

Same as new but also allows to set multicore options.

fn from_array_with_options(is_complex: bool, domain: DataVectorDomain, data: &[f64], options: MultiCoreSettings) -> Self

Same as from_array but also allows to set multicore options.

fn from_array_no_copy_with_options(is_complex: bool, domain: DataVectorDomain, data: Vec<f64>, options: MultiCoreSettings) -> Self

Same as from_array_no_copy but also allows to set multicore options.

fn from_array_with_delta_and_options(is_complex: bool, domain: DataVectorDomain, data: &[f64], delta: f64, options: MultiCoreSettings) -> Self

Same as from_array_with_delta but also allows to set multicore options.

fn from_array_no_copy_with_delta_and_options(is_complex: bool, domain: DataVectorDomain, data: Vec<f64>, delta: f64, options: MultiCoreSettings) -> Self

Same as from_array_no_copy_with_delta but also allows to set multicore options.

fn new(is_complex: bool, domain: DataVectorDomain, init_value: f64, length: usize, delta: f64) -> Self

Creates a new generic data vector from the given arguments.

fn from_array(is_complex: bool, domain: DataVectorDomain, data: &[f64]) -> Self

Creates a new generic data vector from the given arguments.

fn from_array_no_copy(is_complex: bool, domain: DataVectorDomain, data: Vec<f64>) -> Self

Creates a new generic data vector from the given arguments.

fn from_array_with_delta(is_complex: bool, domain: DataVectorDomain, data: &[f64], delta: f64) -> Self

Creates a new generic data vector from the given arguments.

fn from_array_no_copy_with_delta(is_complex: bool, domain: DataVectorDomain, data: Vec<f64>, delta: f64) -> Self

Creates a new generic data vector from the given arguments.

impl GenericDataVector<f32>

fn perform_operations(self, operations: &[Operation<f32>]) -> Self

Perform a set of operations on the given vector. Warning: Highly unstable and not even fully implemented right now.

With this approach we change how we operate on vectors. If you perform M operations on a vector with the length N you iterate wit hall other methods like this:

// pseudocode:
// for m in M:
//  for n in N:
//    execute m on n

with this method the pattern is changed slighly:

// pseudocode:
// for n in N:
//  for m in M:
//    execute m on n

Both variants have the same complexity however the second one is benificial since we have increased locality this way. This should help us by making better use of registers and CPU buffers. This might also help since for large data we might have the chance in future to move the data to a GPU, run all operations and get the result back. In this case the GPU is fast for many operations but the roundtrips on the bus should be minimized to keep the speed advantage.