Trait basic_dsp::ComplexIndex

pub trait ComplexIndex<Idx> {
    type Output: ?Sized;
    fn complex(&self, index: Idx) -> &Self::Output;
}

Like std::ops::Index but with a different method name so that it can be used to implement an additional range accessor for complex data.

Note if indexers will return an empty array in case the vector isn't complex.

Associated Types

type Output: ?Sized

Required Methods

fn complex(&self, index: Idx) -> &Self::Output

The method for complex indexing