Trait basic_dsp::ElementaryOps

pub trait ElementaryOps<A, T, N, D> where
    A: GetMetaData<T, N, D>,
    D: Domain,
    N: NumberSpace,
    T: RealNumber, {
    fn add(&mut self, summand: &A) -> Result<(), ErrorReason>;

    fn sub(&mut self, subtrahend: &A) -> Result<(), ErrorReason>;

    fn mul(&mut self, factor: &A) -> Result<(), ErrorReason>;

    fn div(&mut self, divisor: &A) -> Result<(), ErrorReason>;
}

Elementary algebra on types: addition, subtraction, multiplication and division

Required methods

fn add(&mut self, summand: &A) -> Result<(), ErrorReason>

Calculates the sum of self + summand. It consumes self and returns the result.

Failures

TransRes may report the following ErrorReason members:

VectorsMustHaveTheSameSize: self and summand must have the same size

VectorMetaDataMustAgree: self and summand must be in the same domain and number space

Example

use basic_dsp_vector::*;
let mut vector1 = vec!(1.0, 2.0).to_real_time_vec();
let vector2 = vec!(10.0, 11.0).to_real_time_vec();
vector1.add(&vector2).expect("Ignoring error handling in examples");
assert_eq!([11.0, 13.0], vector1[0..]);

fn sub(&mut self, subtrahend: &A) -> Result<(), ErrorReason>

Calculates the difference of self - subtrahend. It consumes self and returns the result.

Failures

TransRes may report the following ErrorReason members:

VectorsMustHaveTheSameSize: self and subtrahend must have the same size

VectorMetaDataMustAgree: self and subtrahend must be in the same domain and number space

Example

use basic_dsp_vector::*;
let mut vector1 = vec!(1.0, 2.0).to_real_time_vec();
let vector2 = vec!(10.0, 11.0).to_real_time_vec();
vector1.sub(&vector2).expect("Ignoring error handling in examples");
assert_eq!([-9.0, -9.0], vector1[0..]);

fn mul(&mut self, factor: &A) -> Result<(), ErrorReason>

Calculates the product of self * factor. It consumes self and returns the result.

Failures

TransRes may report the following ErrorReason members:

VectorsMustHaveTheSameSize: self and factor must have the same size

VectorMetaDataMustAgree: self and factor must be in the same domain and number space

Example

use basic_dsp_vector::*;
let mut vector1 = vec!(1.0, 2.0).to_real_time_vec();
let vector2 = vec!(10.0, 11.0).to_real_time_vec();
vector1.mul(&vector2).expect("Ignoring error handling in examples");
assert_eq!([10.0, 22.0], vector1[0..]);

fn div(&mut self, divisor: &A) -> Result<(), ErrorReason>

Calculates the quotient of self / summand. It consumes self and returns the result.

Failures

TransRes may report the following ErrorReason members:

VectorsMustHaveTheSameSize: self and divisor must have the same size

VectorMetaDataMustAgree: self and divisor must be in the same domain and number space

Example

use basic_dsp_vector::*;
let mut vector1 = vec!(10.0, 22.0).to_real_time_vec();
let vector2 = vec!(2.0, 11.0).to_real_time_vec();
vector1.div(&vector2).expect("Ignoring error handling in examples");
assert_eq!([5.0, 2.0], vector1[0..]);

Loading content...