Trait basic_dsp::InterleaveToVector

pub trait InterleaveToVector<T>: ToSlice<T> where
    T: RealNumber, {
    fn interleave_to_complex_time_vec(
        &self, 
        other: &Self
    ) -> Result<DspVec<Vec<T>, T, Complex, Time>, ErrorReason>;

    fn interleave_to_complex_freq_vec(
        &self, 
        other: &Self
    ) -> Result<DspVec<Vec<T>, T, Complex, Freq>, ErrorReason>;
}

Conversion from two instances of a generic data type into a dsp vector with complex data.

Required methods

fn interleave_to_complex_time_vec(
&self,
other: &Self
) -> Result<DspVec<Vec<T>, T, Complex, Time>, ErrorReason>

Create a new vector in real number space and time domain. delta can be changed after construction with a call of set_delta.

fn interleave_to_complex_freq_vec(
&self,
other: &Self
) -> Result<DspVec<Vec<T>, T, Complex, Freq>, ErrorReason>

Create a new vector in real number space and frequency domain. delta can be changed after construction with a call of set_delta.

Loading content...