Trait basic_dsp::MapAggregateOps

pub trait MapAggregateOps<T, R> where
    R: Send, {
    type Output;
    fn map_aggregate<'a, A, FMap, FAggr>(
        &self, 
        argument: A, 
        map: &FMap, 
        aggregate: &FAggr
    ) -> Self::Output
    where
        A: Sync + Copy + Send,
        FAggr: Fn(R, R) -> R + 'a + Sync + Send,
        FMap: Fn(T, usize, A) -> R + 'a + Sync;
}

Operations which allow to iterate over the vector and to derive results.

Associated Types

type Output

Required Methods

fn map_aggregate<'a, A, FMap, FAggr>(
&self,
argument: A,
map: &FMap,
aggregate: &FAggr
) -> Self::Output where
A: Sync + Copy + Send,
FAggr: Fn(R, R) -> R + 'a + Sync + Send,
FMap: Fn(T, usize, A) -> R + 'a + Sync,

Transforms all vector elements using the function map and then aggregates all the results with aggregate. aggregate must be a commutativity and associativity; that's because there is no guarantee that the numbers will be aggregated in any deterministic order.