Trait basic_dsp::MetaData

pub trait MetaData {
    fn domain(&self) -> DataDomain;

    fn is_complex(&self) -> bool;
}

A trait which provides information about number space and domain.

Required methods

fn domain(&self) -> DataDomain

The domain in which the data vector resides. Basically specifies the x-axis and the type of operations which are valid on this vector.

The domain can be changed using the RededicateOps trait.

fn is_complex(&self) -> bool

Indicates whether the vector contains complex data. This also specifies the type of operations which are valid on this vector.

The number space can be changed using the RededicateOps trait.

Loading content...