Trait basic_dsp::ResizeBufferedOps

pub trait ResizeBufferedOps<S, T> where
    S: ToSliceMut<T>,
    T: RealNumber, {
    fn resize_b<B>(
        &mut self, 
        buffer: &mut B, 
        len: usize
    ) -> Result<(), ErrorReason>
    where
        B: for<'a> Buffer<'a, S, T>;
}

Operations to resize a data type.

Required methods

fn resize_b<B>(&mut self, buffer: &mut B, len: usize) -> Result<(), ErrorReason> where
B: for<'a> Buffer<'a, S, T>,

Changes self.len(). If self.is_complex() is true then len must be an even number. len > self.alloc_len() is only possible if the underlying storage or the buffer supports resizing.

Loading content...