Trait basic_dsp::ResizeOps

pub trait ResizeOps {
    fn resize(&mut self, len: usize) -> Result<(), ErrorReason>;
}

Operations to resize a data type.

Required methods

fn resize(&mut self, len: usize) -> Result<(), ErrorReason>

Changes self.len(). If self.is_complex() is true then len must be an even number. len > self.alloc_len() is only possible if the underlying storage supports resizing.

Loading content...