Trait basic_dsp::SumOps

pub trait SumOps<T> {
    fn sum(&self) -> T;

    fn sum_sq(&self) -> T;
}

Offers operations to calculate the sum or the sum of squares.

Required methods

fn sum(&self) -> T

Calculates the sum of the data contained in the vector.

Example

use basic_dsp_vector::*;
let vector = vec!(Complex32::new(1.0, 2.0), Complex32::new(3.0, 4.0), Complex32::new(5.0, 6.0)).to_complex_time_vec();
let result = vector.sum();
assert_eq!(result, Complex32::new(9.0, 12.0));
}

fn sum_sq(&self) -> T

Calculates the sum of the squared data contained in the vector.

Example

use basic_dsp_vector::*;
let vector = vec!(Complex64::new(1.0, 2.0), Complex64::new(3.0, 4.0), Complex64::new(5.0, 6.0)).to_complex_time_vec();
let result = vector.sum_sq();
assert_eq!(result, Complex64::new(-21.0, 88.0));
}

Loading content...