Trait basic_dsp::SymmetricFrequencyDomainOperations

pub trait SymmetricFrequencyDomainOperations<T>: DataVector<T> where T: RealNumber {
    type RealTimePartner;
    fn plain_sifft(self) -> VecResult<Self::RealTimePartner>;
    fn sifft(self) -> VecResult<Self::RealTimePartner>;
    fn windowed_sifft(self, window: &WindowFunction<T>) -> VecResult<Self::RealTimePartner>;
}

Defines all operations which are valid on DataVectors containing frequency domain data and the data is assumed to half of complex conjugate symmetric spectrum round 0 Hz where the 0 Hz element itself is real.

Failures

All operations in this trait fail with VectorMustBeInFrquencyDomain if the vector isn't in frequency domain or with VectorMustBeConjSymmetric if the first element (0Hz) isn't real.

Associated Types

type RealTimePartner

Required Methods

fn plain_sifft(self) -> VecResult<Self::RealTimePartner>

Performs a Symmetric Inverse Fast Fourier Transformation under the assumption that self contains half of a symmetric spectrum starting from 0 Hz. This assumption isn't verified and no error is raised if the spectrum isn't symmetric. The reason for this is that there is no robust verification possible.

The argument indicates whether the resulting real vector should have 2*N or 2*N-1 points.

This version of the IFFT neither applies a window nor does it scale the vector.

fn sifft(self) -> VecResult<Self::RealTimePartner>

Performs a Symmetric Inverse Fast Fourier Transformation under the assumption that self contains half of a symmetric spectrum starting from 0 Hz. This assumption isn't verified and no error is raised if the spectrum isn't symmetric. The reason for this is that there is no robust verification possible.

The argument indicates whether the resulting real vector should have 2*N or 2*N-1 points.

fn windowed_sifft(self, window: &WindowFunction<T>) -> VecResult<Self::RealTimePartner>

Performs a Symmetric Inverse Fast Fourier Transformation (SIFFT) and removes the FFT window. The SIFFT is performed under the assumption that self contains half of a symmetric spectrum starting from 0 Hz. This assumption isn't verified and no error is raised if the spectrum isn't symmetric. The reason for this is that there is no robust verification possible.

The argument indicates whether the resulting real vector should have 2*N or 2*N-1 points.