Trait basic_dsp::ToRealVector

pub trait ToRealVector<T>: ToSlice<T> where
    T: RealNumber, {
    fn to_real_time_vec(self) -> DspVec<Self, T, Real, Time>;

    fn to_real_freq_vec(self) -> DspVec<Self, T, Real, Freq>;
}

Conversion from a generic data type into a dsp vector with real data.

Required methods

fn to_real_time_vec(self) -> DspVec<Self, T, Real, Time>

Create a new vector in real number space and time domain. delta can be changed after construction with a call of set_delta.

fn to_real_freq_vec(self) -> DspVec<Self, T, Real, Freq>

Create a new vector in real number space and frequency domain. delta can be changed after construction with a call of set_delta.

Loading content...