Module basic_dsp::interop_facade::facade64

Functions for 64bit floating point number based vectors. Please refer to the other chapters of the help for documentation of the functions.

Functions

abs64

acos64

acosh64

add_smaller_vector64

add_vector64

apply_custom_window64
Creates a window from the function window and the void pointer window_data. The window_data pointer is passed to the window function at every call and can be used to store parameters.

apply_window64
window argument is translated to:

asin64

asinh64

atan64

atanh64

clone64

complex_conj64

complex_divide64

complex_dot_product64

complex_offset64

complex_scale64

complex_statistics64

complex_statistics_splitted64

cos64

cosh64

cum_sum64

delete_vector64

diff64

diff_with_start64

divide_smaller_vector64

divide_vector64

exp_base64

expn64

fft64

get_allocated_len64

get_delta64

get_domain64
Returns the vector domain as integer:

get_imag64

get_len64

get_mag_phase64

get_magnitude64

get_phase64

get_points64

get_real64

get_real_imag64

get_value64

ifft64

is_complex64

log_base64

logn64

magnitude64

magnitude_squared64

merge64

mirror64

multiply_complex_exponential64

multiply_smaller_vector64

multiply_vector64

new64

override_data64

phase64

plain_fft64

plain_ifft64

plain_sifft64

power64

real_dot_product64

real_offset64

real_scale64

real_statistics64

real_statistics_splitted64

root64

set_mag_phase64

set_real_imag64

set_value64

sifft64

sin64

sinh64

split_into64

sqrt64

square64

subtract_smaller_vector64

subtract_vector64

swap_halves64

tan64

tanh64

to_complex64

to_imag64

to_real64

unapply_custom_window64
See apply_custom_window64 for a description of the window and window_data parameter.

unapply_window64
See apply_window64 for a description of the window parameter.

unwrap64

windowed_custom_fft64
See apply_custom_window64 for a description of the window and window_data parameter.

windowed_custom_ifft64
See apply_custom_window64 for a description of the window and window_data parameter.

windowed_custom_sifft64
See apply_custom_window64 for a description of the window and window_data parameter.

windowed_fft64
See apply_window64 for a description of the window parameter.

windowed_ifft64
See apply_window64 for a description of the window parameter.

windowed_sifft64
See apply_window64 for a description of the window parameter.

wrap64

zero_interleave64

zero_pad64
padding_option argument is translated to: Returns the vector domain as integer: