Trait basic_dsp::ComplexIndexMut

pub trait ComplexIndexMut<Idx>: ComplexIndex<Idx> {
    fn complex_mut(&mut self, index: Idx) -> &mut Self::Output;
}

Like std::ops::IndexMut but with a different method name so that it can be used to implement a additional range accessor for complex data.

Note if indexers will return an empty array in case the vector isn't complex.

Required methods

fn complex_mut(&mut self, index: Idx) -> &mut Self::Output

The method for complex indexing

Loading content...