Trait basic_dsp::OffsetOps

pub trait OffsetOps<T> {
    fn offset(&mut self, offset: T);
}

An operation which adds a constant to each vector element

Required Methods

fn offset(&mut self, offset: T)

Adds a scalar to each vector element.

Failures

self.len() to 0 if the vector isn't in the complex number space but factor is complex.

Example

use basic_dsp_vector::*;
let mut vector = vec!(1.0, 2.0).to_real_time_vec();
vector.offset(2.0);
assert_eq!([3.0, 4.0], vector[0..]);